anthropic[patch]: Proper handling of multi tool result messages
