+++ libs/langchain-anthropic/src/tests/chat_models-tools.int.test.ts
@@ -440,3 +440,57 @@ test("llm token callbacks can handle tool calls", async () => {
   if (!args) return;
   expect(args).toEqual(JSON.parse(tokens));
 });
+
+test("Can process dual tool use and tool result", async () => {
+  const weatherTool = tool((_) => "no-op", {
+    name: "get_weather",
+    description: "Get the weather of a specific location.",
+    schema: z.object({
+      location: z.string().describe("The name of city to get the weather for."),
+    }),
+  });
+  const calculatorTool = tool((_) => "no-op", {
+    name: "calculator",
+    description: "Execute a math expression.",
+    schema: z.object({
+      expression: z.string().describe("The math expression to calculate."),
+    }),
+  });
+
+  const modelWithTools = model.bindTools([weatherTool, calculatorTool]);
+  const messageHistory = [
+    new HumanMessage("What is the weather in SF? Also, what is 2 + 2?"),
+    new AIMessage({
+      content: "",
+      tool_calls: [
+        {
+          name: "get_weather",
+          id: "weather_call",
+          args: {
+            location: "SF",
+          },
+        },
+        {
+          name: "calculator",
+          id: "calculator_call",
+          args: {
+            expression: "2 + 2",
+          },
+        },
+      ],
+    }),
+    new ToolMessage({
+      name: "get_weather",
+      tool_call_id: "weather_call",
+      content: "It is currently 24 degrees with hail in San Francisco.",
+    }),
+    new ToolMessage({
+      name: "calculator",
+      tool_call_id: "calculator_call",
+      content: "2 + 2 = 4",
+    }),
+  ];
+
+  const result = await modelWithTools.invoke(messageHistory);
+  console.dir(result, { depth: null });
+});
