+++ libs/langchain-anthropic/src/types.ts
@@ -1,3 +1,5 @@
+import Anthropic from "@anthropic-ai/sdk";
+
 export type AnthropicToolResponse = {
   type: "tool_use";
   id: string;
@@ -5,3 +7,13 @@ export type AnthropicToolResponse = {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   input: Record<string, any>;
 };
+export type AnthropicMessageParam = Anthropic.MessageParam;
+export type AnthropicMessageResponse =
+  | Anthropic.ContentBlock
+  | AnthropicToolResponse;
+export type AnthropicMessageCreateParams =
+  Anthropic.MessageCreateParamsNonStreaming;
+export type AnthropicStreamingMessageCreateParams =
+  Anthropic.MessageCreateParamsStreaming;
+export type AnthropicMessageStreamEvent = Anthropic.MessageStreamEvent;
+export type AnthropicRequestOptions = Anthropic.RequestOptions;
